Correct calendar getter contracts and refine draw recovery evidence

+++ src/promoted/code1_0011.c
@@ -67,9 +67,9 @@ extern s32 func_00106c30(s16 arg0, s16 arg1);
 extern s32 func_00106c80(s16 arg0);
 extern s16 func_00106cd0(s16 arg0, s16 arg1);
 extern u8 *func_0010d620(s16 arg0);
-extern s64 func_00123ae0(void);
-extern s64 func_00123b10(void);
-extern s64 func_00123b40(void);
+extern s8 func_00123ae0(void);
+extern s16 func_00123b10(void);
+extern s16 func_00123b40(void);
 extern u8 D_005E4F70[];
 extern s8 D_005E47F0[];
 static inline s64 p4_0011_signext16(s16 arg0)
